Invoke virtualenv for fabric commands.
Fabric invoke commands as '/bin/bash -l -c' which fires off a login session
but non-interactive. It means that ~/.bashrc will not be read and thus
virtualenv will not be activated. To get around this, virtualenv should be
activated in ~/.profile as well once this file will be read by bash in a
session described as above.